requirements.txt
about-time==4.2.1
accelerate==0.28.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alive-progress==3.1.5
altair==5.2.0
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
blessed==1.20.0
blinker==1.7.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
cma==3.2.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
config==0.5.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.27.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
editor==1.6.6
einops==0.7.0
emoji==2.10.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.7
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.2
googleapis-common-protos==1.63.0
gpt4all==2.2.1.post1
gradio==4.24.0
gradio_client==0.14.0
grapheme==0.6.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.3.0
inquirer==3.2.4
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jinxed==1.2.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jstyleson==0.0.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
lancedb==0.6.5
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.24
llama_cpp_python==0.2.56
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
nncf @ git+https://github.com/openvinotoolkit/nncf.git@fa8b702933986a65cb7981d1df2bc05ce80b356a
notebook-utils==0.2.0
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.0
onnxruntime==1.17.1
openai==1.13.3
opencv-python==4.9.0.80
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
openvino==2024.0.0
openvino-nightly==2024.1.0.dev20240328
openvino-telemetry==2023.2.1
#openvino-tensorflow==2.3.0
optimum==1.18.0
optimum-intel @ git+https://github.com/huggingface/optimum-intel.git@447ef502b7f4fd089d43c24de42394d9ac8aac1d
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pdfminer.six==20231228
peewee==3.17.1
pillow==10.2.0
platformdirs==4.2.0
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
pylance==0.10.5
pymoo==0.6.1.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
ratelimiter==1.2.0.post0
readchar==4.0.6
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.4
runs==1.2.2
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==2.6.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
streamlit==1.32.1
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tavily-python==0.3.1
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.5.2
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.38.2
transformers-stream-generator==0.0.5
typer==0.9.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.12.6
unstructured-client==0.18.0
urllib3==1.26.18
uvicorn==0.28.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
widgetsnbextension==4.0.10
wrapt==1.16.0
xmod==1.8.1
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.37
zipp==3.18.0
